+++ src/datastore_pandas/__init__.py
@@ -46,7 +46,7 @@ from datastore_pandas.types import (
     TimestampType,
 )
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 
 __all__ = [
     "ArrayType",